+++ Toggly.FeatureManagement.Next/nextjs-toggly-client/src/context.tsx
@@ -284,7 +284,10 @@ function TogglyProviderOwner({
     <T = unknown>(
       featureKey: string,
       isT?: (v: unknown) => v is T,
-    ): T | null => client.getVariantValue(featureKey, isT),
+    ): T | null =>
+      isT === undefined
+        ? client.getVariantValue(featureKey)
+        : client.getVariantValue(featureKey, isT),
     [client]
   )